requirements.txt
Django==4.0.2
djangorestframework==3.13.1
drf-yasg==1.20.0
# django-cors-headers==3.11.0
psycopg2==2.9.3
# psycopg2-binary==2.9.3




